fix: make BroadcastReceiver safe from 3rd-party calls
ACTION_BOOT_COMPLETED can only be sent by the system; however, we are
not checking for that specific intent in our BroadcastReceiver class.
This allows any third-party app to call our BR class and execute the
code, which is not what we want. To fix this, check the intent received
before executing any code.

Further information here: https://stackoverflow.com/a/52628842