fix: force resizable ratios for low SDK
Because we are targeting Android SDK 22, not setting this attribute
causes Android to set a conservative ratio for our app size. There is a
way to override it in system settings, but we should just let the system
know that our app is fine to resize to save users the hassle.